clean ups, bundle xray function into separate file
